requirements.txt
certifi==2020.12.5
cycler==0.10.0
Cython==0.29.22
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
six==1.15.0
tqdm==4.57.0
